Compensates for image path resoution betwen LaTeX and EPUB
Pandoc seems to generate the paths for images differently depending on
which type of output you are generating. This cats all the text files
together in a temp directory and copies the images into the same
directory, so that the paths can be specified in the source without any
parent directories.